chore: split facility_report api/service out of report files
