+++ bc_obps/reporting/api/__init__.py
@@ -4,3 +4,4 @@ from .reports import start_report
 from .build_form_schema import build_form_schema
 from .operations import get_dashboard_operations_list
 from .activity_data import get_initial_activity_data
+from .facility_report import get_facility_report_form_data
