+++ bc_obps/reporting/api/reports.py
@@ -1,5 +1,4 @@
-from typing import Literal, Tuple, Optional
-from uuid import UUID
+from typing import Literal, Tuple
 from django.http import HttpRequest
 from registration.decorators import handle_http_errors
 from reporting.constants import EMISSIONS_REPORT_TAGS
@@ -12,7 +11,6 @@ from reporting.schema.report_operation import ReportOperationOut, ReportOperatio
 from reporting.schema.reporting_year import ReportingYearOut
 from .router import router
 from ..models import ReportingYear
-from ..schema.facility_report import FacilityReportOut, FacilityReportIn
 
 
 @router.post(
@@ -67,59 +65,3 @@ def save_report(
 @handle_http_errors()
 def get_reporting_year(request: HttpRequest) -> Tuple[Literal[200], ReportingYear]:
     return 200, ReportingYearService.get_current_reporting_year()
-
-
-@router.get(
-    "/report-version/{version_id}/facility-report/{facility_id}",
-    response={200: FacilityReportOut, 404: Message, 400: Message, 500: Message},
-    tags=EMISSIONS_REPORT_TAGS,
-    description="""Takes `version_id` (primary key of the ReportVersion model) and `facility_id` to return a single matching `facility_report` object.
-    Includes the associated activity IDs if found; otherwise, returns an error message if not found or in case of other issues.""",
-)
-@handle_http_errors()
-def get_facility_report_form_data(
-    request: HttpRequest, version_id: int, facility_id: UUID
-) -> Tuple[Literal[200], Optional[FacilityReportOut]]:
-    facility_report = ReportService.get_facility_report_by_version_and_id(version_id, facility_id)
-    activity_ids = ReportService.get_activity_ids_for_facility(version_id, facility_id) or []
-    return 200, ReportService.get_facility_report_form_data(facility_report, activity_ids)
-
-
-@router.post(
-    "/report-version/{version_id}/facility-report/{facility_id}",
-    response={201: FacilityReportOut, custom_codes_4xx: Message},
-    tags=EMISSIONS_REPORT_TAGS,
-    description="""Updates the report facility details by version_id and facility_id. The request body should include
-    fields to be updated, such as facility name, type, BC GHG ID, activities, and products. Returns the updated report
-    facility object or an error message if the update fails.""",
-)
-@handle_http_errors()
-def save_facility_report(
-    request: HttpRequest, version_id: int, facility_id: UUID, payload: FacilityReportIn
-) -> Tuple[Literal[201], FacilityReportOut]:
-    """
-    Save or update a report facility and its related activities.
-
-    Args:
-        request (HttpRequest): The HTTP request object.
-        version_id (int): The ID of the report version.
-        facility_id (int): The ID of the facility.
-        payload (FacilityReportIn): The input data for the report facility.
-
-    Returns:
-        Tuple: HTTP status code and the response data or an error message.
-    """
-    # Fetch the existing facility report or create a new one
-    facility_report = ReportService.save_facility_report(version_id, facility_id, payload)
-
-    # Prepare the response data
-    response_data = FacilityReportOut(
-        id=facility_id,
-        report_version_id=facility_report.report_version.id,
-        facility_name=facility_report.facility_name,
-        facility_type=facility_report.facility_type,
-        facility_bcghgid=facility_report.facility_bcghgid,
-        activities=list(facility_report.activities.values_list('id', flat=True)),
-        products=list(facility_report.products.values_list('id', flat=True)) or [],
-    )
-    return 201, response_data
